fix 500 for non-admin members: SELECT 1 is INT4, not INT8
can_view_system decoded Postgres' SELECT 1 as i64, so every system and cluster
endpoint returned 500 for workspace viewers/editors while admins short-circuited
on read-all and saw nothing wrong. Add scripts/check-viewer-system-access.sh,
which fails against the old code and passes now.

+++ crates/hub/src/web/mod.rs
@@ -28,7 +28,8 @@ pub async fn can_view_system(
     if user.can_read_all() {
         return Ok(true);
     }
-    let row: Option<(i64,)> = sqlx::query_as(
+    // `SELECT 1` is INT4 in Postgres — decoding it as i64 fails at runtime (500).
+    let row: Option<(i32,)> = sqlx::query_as(
         "SELECT 1 FROM systems s \
          JOIN memberships m ON m.workspace_id = s.workspace_id \
          WHERE s.id = $1 AND m.user_id = $2",
